snapshot sperator
